Cleaned Files. Fixed Hitbox
Cleaned up the white space in most of the files, fixed some of the recipes to have actual identifiers rather than just random variables. Fixed the hitbox on the drones. Made the shell produce 2 of its self each time and take half the resources so the resources used are the same. Fixed the Drone Model but then broke it again.